requirements.txt
python-dotenv~=1.0.1
caldav~=1.3.9
icalendar~=5.0.13